chore: split checks into rules
