+++ src/crs_linter/rules/__init__.py
@@ -1,0 +1,17 @@
+# import all checks
+
+from . import (
+    approved_tags,
+    capture,
+    crs_tag,
+    deprecated,
+    duplicated,
+    ignore_case,
+    indentation,
+    lowercase_ignorecase,
+    ordered_actions,
+    pl_consistency,
+    pl_tags,
+    variables_usage,
+    version
+)
